requirements.txt
python-telegram-bot==21.9
google-genai>=1.5.0
duckduckgo-search>=6.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
dotenv